+++ Css/WebScrap.Css/Tokens/CssToken.cs
@@ -1,8 +1,0 @@
-namespace WebScrap.Css.Tokens;
-
-internal record struct CssToken(
-    ReadOnlyMemory<char> Tag,
-    char? ChildSelector, 
-    ILookup<string, string> Attributes)
-{
-}