+++ WebScrap.Common/Processors/IProcessorListenerExtensions.cs
@@ -1,0 +1,13 @@
+using WebScrap.Common.Tags;
+
+namespace WebScrap.Common.Processors;
+
+internal static class IProcessorListenerExtensions
+{
+    internal static TListener Get<TListener>(
+        this IEnumerable<IProcessorListener> listeners)
+            where TListener : IProcessorListener
+        => listeners.OfType<TListener>().First();
+
+    
+}