+++ Changelog.md
@@ -1,2 +1,3 @@
-v0.2 - Improved `HtmlExtraction` feature with class and id selectors.
+v0.3 - Added space selector support.
+v0.2 - Added id and class selector support.
 v0.1 - Added `HtmlExtraction` feature. Plain tags.